update hibernate, Jpa

+++ src/main/java/com/example/SpringMVC/dao/LectureRepository.java
@@ -1,10 +1,8 @@
 package com.example.SpringMVC.dao;
 
-
-
 import com.example.SpringMVC.model.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LectureRepository extends JpaRepository<Lecture, Integer> {
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
 
 }
